LRQA-42795 Rework the GitBranch hierarchy to be based off of GitRef
