+++ modules/test/jenkins-results-parser/src/main/java/com/liferay/jenkins/results/parser/LocalGitBranch.java
@@ -19,7 +19,7 @@ import java.io.File;
 /**
  * @author Michael Hashimoto
  */
-public class LocalGitBranch extends BaseGitBranch {
+public class LocalGitBranch extends BaseGitRef {
 
 	public File getDirectory() {
 		LocalRepository localRepository = getLocalRepository();
