+++ modules/test/jenkins-results-parser/src/main/java/com/liferay/jenkins/results/parser/RemoteGitRef.java
@@ -1,0 +1,69 @@
+/**
+ * Copyright (c) 2000-present Liferay, Inc. All rights reserved.
+ *
+ * This library is free software; you can redistribute it and/or modify it under
+ * the terms of the GNU Lesser General Public License as published by the Free
+ * Software Foundation; either version 2.1 of the License, or (at your option)
+ * any later version.
+ *
+ * This library is distributed in the hope that it will be useful, but WITHOUT
+ * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
+ * FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more
+ * details.
+ */
+
+package com.liferay.jenkins.results.parser;
+
+/**
+ * @author Michael Hashimoto
+ */
+public class RemoteGitRef extends BaseGitRef implements Comparable<RemoteGitRef> {
+
+	@Override
+	public int compareTo(RemoteGitRef o) {
+		String name = getName();
+
+		return name.compareTo(o.getName());
+	}
+
+	public RemoteRepository getRemoteRepository() {
+		return _remoteRepository;
+	}
+
+	public String getUsername() {
+		RemoteRepository remoteRepository = getRemoteRepository();
+
+		return remoteRepository.getUsername();
+	}
+
+	@Override
+	public String toString() {
+		RemoteRepository remoteRepository = getRemoteRepository();
+
+		StringBuilder sb = new StringBuilder();
+
+		sb.append(remoteRepository.getRemoteURL());
+		sb.append(" (");
+		sb.append(getName());
+		sb.append(" - ");
+		sb.append(getSHA());
+		sb.append(")");
+
+		return sb.toString();
+	}
+
+	protected RemoteGitRef(
+		RemoteRepository remoteRepository, String name, String sha) {
+
+		super(name, sha);
+
+		if (remoteRepository == null) {
+			throw new IllegalArgumentException("Remote repository is null");
+		}
+
+		_remoteRepository = remoteRepository;
+	}
+
+	private final RemoteRepository _remoteRepository;
+
+}
